Add files via upload
Implement flow of a factory pattern for the creation of a car, in which there are different objects like seats, engine, doors, multimedia, suspension, electrical system etc. Ensuring only a single object of the Car Factory is ever available, keep track of cars produced by the factory and provide an interface to get a list of all those cars and their counts.
Factory should help make a car with custom requirements for all mentioned components, with sensible defaults set.